chore: rename test files to remove redundant name fragment
